requirements.txt
 streamlit==1.32.0
 openai-whisper